+++ packages/client/src/RpcProviderSource.ts
@@ -3,11 +3,12 @@ import { FallbackProvider, FetchRequest, JsonRpcProvider } from 'ethers'
 import { Lifecycle, inject, scoped } from 'tsyringe'
 import { ConfigInjectionToken, StrictStreamrClientConfig } from './Config'
 import { LoggingJsonRpcProvider } from './utils/LoggingJsonRpcProvider'
+import { config as CHAIN_CONFIG } from '@streamr/config'
 
 export const QUORUM = 2
 
 function isDevChain(config: Pick<StrictStreamrClientConfig, 'contracts'>): boolean {
-    return config.contracts.streamRegistryChainRPCs?.name === 'dev2'
+    return config.contracts.ethereumNetwork.chainId === CHAIN_CONFIG.dev2.id
 }
 
 @scoped(Lifecycle.ContainerScoped)
@@ -23,23 +24,17 @@ export class RpcProviderSource {
         if (this.provider === undefined) {
             // eslint-disable-next-line no-underscore-dangle
             const timeout = this.config._timeouts.jsonRpcTimeout
-            const providers = this.config.contracts.streamRegistryChainRPCs.rpcs.map((c) => {
+            const providers = this.config.contracts.rpcs.map((c) => {
                 const fetchRequest = new FetchRequest(c.url)
                 fetchRequest.timeout = timeout
-                return new LoggingJsonRpcProvider(fetchRequest, {
-                    chainId: this.config.contracts.streamRegistryChainRPCs.chainId,
-                    name: this.config.contracts.streamRegistryChainRPCs.name
-                }, {
+                return new LoggingJsonRpcProvider(fetchRequest, this.config.contracts.ethereumNetwork.chainId, {
                     staticNetwork: true,
                     batchStallTime: isDevChain(this.config) ? 0 : undefined, // Don't batch requests, send them immediately
                     cacheTimeout: isDevChain(this.config) ? -1 : undefined   // Do not employ result caching
                 })
             })
-            this.provider = new FallbackProvider(providers, {
-                chainId: this.config.contracts.streamRegistryChainRPCs.chainId,
-                name: this.config.contracts.streamRegistryChainRPCs.name
-            }, {
-                quorum: Math.min(QUORUM, this.config.contracts.streamRegistryChainRPCs?.rpcs.length),
+            this.provider = new FallbackProvider(providers, this.config.contracts.ethereumNetwork.chainId, {
+                quorum: Math.min(QUORUM, this.config.contracts.rpcs.length),
                 cacheTimeout: isDevChain(this.config) ? -1 : undefined   // Do not employ result caching
             })
         }
@@ -48,13 +43,10 @@ export class RpcProviderSource {
 
     // TODO reduce copy-paste?
     getEventProviders(): Provider[] {
-        return this.config.contracts.streamRegistryChainRPCs.rpcs.map((c) => {
+        return this.config.contracts.rpcs.map((c) => {
             const f = new FetchRequest(c.url)
             f.retryFunc = async () => false  // TODO use this line also for providers in the fallbackprovider?
-            return new JsonRpcProvider(f, {
-                chainId: this.config.contracts.streamRegistryChainRPCs.chainId,
-                name: this.config.contracts.streamRegistryChainRPCs.name
-            }, {
+            return new JsonRpcProvider(f, this.config.contracts.ethereumNetwork.chainId, {
                 staticNetwork: true
             })
         })